refactor: error messages
