+++ src/hooks/index.js
@@ -1,4 +1,3 @@
-export { useCommonErrorMessages } from './useCommonErrorMessages';
 export { useCurrentConsortium } from './useCurrentConsortium';
 export { useCurrentUserTenantsPermissions } from './useCurrentUserTenantsPermissions';
 export { useEventEmitter } from './useEventEmitter';
