--- src/hooks/useCommonErrorMessages/index.js
@@ -1,1 +1,0 @@
-export { useCommonErrorMessages } from './useCommonErrorMessages';
