+++ src/routes/ConsortiumManager/settings/authorization-roles/AuthorizationRolesViewPage.js
@@ -1,8 +1,15 @@
 import PropTypes from 'prop-types';
-import { useMemo, useState } from 'react';
-import { FormattedMessage } from 'react-intl';
+import {
+  useMemo,
+  useState,
+} from 'react';
+import {
+  FormattedMessage,
+  useIntl,
+} from 'react-intl';
 import { useParams } from 'react-router-dom';
 
+import { useShowCallout } from '@folio/stripes-acq-components';
 import {
   Button,
   Dropdown,
@@ -21,15 +28,17 @@ import {
   useAuthorizationRoles,
 } from '@folio/stripes-authorization-components';
 
+import { handleErrorMessages } from '../../../../utils';
 import { useMemberSelectionContext } from '../../MemberSelectionContext';
 import {
   COLUMN_MAPPING,
   VISIBLE_COLUMNS,
 } from './constants';
 import { getResultsFormatter } from './utils';
-import { useCommonErrorMessages } from '../../../../hooks';
 
 export const AuthorizationRolesViewPage = ({ path }) => {
+  const intl = useIntl();
+  const showCallout = useShowCallout();
   const { id: roleId } = useParams();
   const [searchTerm, setSearchTerm] = useState('');
 
@@ -39,11 +48,12 @@ export const AuthorizationRolesViewPage = ({ path }) => {
     setActiveMember,
   } = useMemberSelectionContext();
 
-  const { handleErrorMessages } = useCommonErrorMessages();
   const { roles, isLoading, onSubmitSearch } = useAuthorizationRoles(
     activeMember,
     {
       onError: ({ response }) => handleErrorMessages({
+        intl,
+        showCallout,
         response,
         messageId: 'ui-consortia-settings.authorizationRoles.errors.loading.data',
       }),
