+++ src/utils/index.js
@@ -6,3 +6,4 @@ export { getModuleName } from './getModuleName';
 export { hydrateSharedRecords } from './hydrateSharedRecords';
 export { isSettingShared } from './isSettingShared';
 export { throwErrorResponse } from './throwErrorResponse';
+export * from './utils';
